fix(qwk): address PR review (Close error, bbsID param, tag dedup, test split)
- qwk.WriteREP: close the zip.Writer explicitly and propagate the
  central-directory flush error instead of dropping it via defer, so a
  finalization failure no longer yields a silently-truncated REP archive.
- qwkservice.ImportREP: drop the redundant bbsID parameter and use the
  service's own bbsID, removing the caller-mismatch foot-gun.
- qwkservice.BuildPacket: build the tag list into a fresh slice (never
  alias/append into the caller's TaggedTags) and dedup tags so a repeated
  tag cannot produce duplicate conferences or last-read updates. Added
  tests for both.
- Clarify the misleading "all accessible areas" doc comment (the fallback
  is not access-filtered).
- Split the 391-line service_test.go into fakestore_test.go,
  service_export_test.go, and service_import_test.go to satisfy the
  <300-line file guideline.
- Design doc: require an atomic check-and-record (SQLite or an enforced
  single-writer model) for REP dedup so retry-safety holds under
  overlapping uploads.

go test ./... (1577 pass), -race clean on new packages, gofmt/go vet clean.
